pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.4.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "feedparser>=6.0.10",
    "pytz>=2023.3",
]
